+++ src/NumHask/Algebra/Field.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
@@ -10,19 +9,27 @@
 module NumHask.Algebra.Field
   ( Field,
     ExpField (..),
+    logBase,
+    sqrt,
     QuotientField (..),
-    UpperBoundedField (..),
-    LowerBoundedField (..),
+    round,
+    ceiling,
+    floor,
+    truncate,
+    infinity,
+    negInfinity,
+    nan,
     TrigField (..),
     half,
   )
 where
 
 import Data.Bool (bool)
-import NumHask.Algebra.Additive (Additive (..), Subtractive (..))
+import NumHask.Algebra.Additive (Additive (..), Subtractive (..), (-))
 import NumHask.Algebra.Multiplicative
-  ( Divisive ((/)),
+  ( Divisive (..),
     Multiplicative (..),
+    (/),
   )
 import NumHask.Algebra.Ring (Distributive, two)
 import NumHask.Data.Integral (Integral, even)
@@ -81,12 +88,22 @@ class
   where
   exp :: a -> a
   log :: a -> a
-  logBase :: a -> a -> a
-  logBase a b = log b / log a
   (**) :: a -> a -> a
   (**) a b = exp (log a * b)
-  sqrt :: a -> a
-  sqrt a = a ** (one / (one + one))
+
+-- | log to the base of
+--
+-- >>> logBase 2 8
+-- 2.9999999999999996
+logBase :: (ExpField a) => a -> a -> a
+logBase a b = log b / log a
+
+-- | square root
+--
+-- >>> sqrt 4
+-- 2.0
+sqrt :: (ExpField a) => a -> a
+sqrt a = a ** (one / (one + one))
 
 instance ExpField P.Double where
   exp = P.exp
@@ -101,9 +118,6 @@ instance ExpField P.Float where
 instance ExpField b => ExpField (a -> b) where
   exp f = exp . f
   log f = log . f
-  logBase f f' a = logBase (f a) (f' a)
-  f ** f' = \a -> f a ** f' a
-  sqrt f = sqrt . f
 
 -- | Conversion from a 'Field' to a 'NumHask.Algebra.Ring'
 --
@@ -114,35 +128,55 @@ instance ExpField b => ExpField (a -> b) where
 class (Field a, Multiplicative b, Additive b) => QuotientField a b where
   properFraction :: a -> (b, a)
 
-  round :: a -> b
-  default round :: (P.Ord a, P.Ord b, Subtractive b, Integral b) => a -> b
-  round x = case properFraction x of
-    (n, r) ->
-      let m = bool (n + one) (n - one) (r P.< zero)
-          half_down = abs' r - (one / (one + one))
-          abs' a
-            | a P.< zero = negate a
-            | P.otherwise = a
-       in case P.compare half_down zero of
-            P.LT -> n
-            P.EQ -> bool m n (even n)
-            P.GT -> m
+-- | round to the nearest integral
+--
+-- Exact ties are managed by rounding down ties if the whole component is even.
+--
+-- >>> round (1.5 :: Double) :: Int
+-- 2
+--
+-- >>> round (2.5 :: Double) :: Int
+-- 2
+round :: (P.Ord a, P.Ord b, QuotientField a b, Subtractive b, Integral b) => a -> b
+round x = case properFraction x of
+  (n, r) ->
+    let m = bool (n + one) (n - one) (r P.< zero)
+        half_down = abs' r - (one / (one + one))
+        abs' a
+          | a P.< zero = negate a
+          | P.otherwise = a
+     in case P.compare half_down zero of
+          P.LT -> n
+          P.EQ -> bool m n (even n)
+          P.GT -> m
 
-  ceiling :: a -> b
-  default ceiling :: (P.Ord a) => a -> b
-  ceiling x = bool n (n + one) (r P.>= zero)
-    where
-      (n, r) = properFraction x
+-- | supply the next upper whole component
+--
+-- >>> ceiling (1.001 :: Double) :: Int
+-- 2
+ceiling :: (P.Ord a, QuotientField a b) => a -> b
+ceiling x = bool n (n + one) (r P.>= zero)
+  where
+    (n, r) = properFraction x
 
-  floor :: a -> b
-  default floor :: (P.Ord a, Subtractive b) => a -> b
-  floor x = bool n (n - one) (r P.< zero)
-    where
-      (n, r) = properFraction x
+-- | supply the previous lower whole component
+--
+-- >>> floor (1.001 :: Double) :: Int
+-- 1
+floor :: (P.Ord a, QuotientField a b, Subtractive b) => a -> b
+floor x = bool n (n - one) (r P.< zero)
+  where
+    (n, r) = properFraction x
 
-  truncate :: a -> b
-  default truncate :: (P.Ord a) => a -> b
-  truncate x = bool (ceiling x) (floor x) (x P.> zero)
+-- | supply the whole component closest to zero
+--
+-- >>> floor (-1.001 :: Double) :: Int
+-- -2
+--
+-- >>> truncate (-1.001 :: Double) :: Int
+-- -1
+truncate :: (P.Ord a, QuotientField a b, Subtractive b) => a -> b
+truncate x = bool (ceiling x) (floor x) (x P.> zero)
 
 instance QuotientField P.Float P.Integer where
   properFraction = P.properFraction
@@ -161,53 +195,26 @@ instance QuotientField b c => QuotientField (a -> b) (a -> c) where
     where
       frac a = properFraction @b @c (f a)
 
-  round f = round . f
-
-  ceiling f = ceiling . f
-
-  floor f = floor . f
-
-  truncate f = truncate . f
-
--- | A bounded field introduces the concepts of infinity and NaN.
+-- | A field introduces the concept of infinity.
 --
 -- > one / zero + infinity == infinity
 -- > infinity + a == infinity
 -- > zero / zero != nan
 --
 -- Note the tricky law that, although nan is assigned to zero/zero, they are never-the-less not equal. A committee decided this.
-class
-  (Field a) =>
-  UpperBoundedField a
-  where
-  infinity :: a
-  infinity = one / zero
+infinity :: (Field a) => a
+infinity = one / zero
 
-  nan :: a
-  nan = zero / zero
+-- |
+--
+-- Note the law:
+-- -- > zero / zero != nan
+nan :: (Field a) => a
+nan = zero / zero
 
-instance UpperBoundedField P.Float
-
-instance UpperBoundedField P.Double
-
-instance UpperBoundedField b => UpperBoundedField (a -> b) where
-  infinity _ = infinity
-  nan _ = nan
-
--- | Negative infinity.
-class
-  (Subtractive a, Field a) =>
-  LowerBoundedField a
-  where
-  negInfinity :: a
-  negInfinity = negate (one / zero)
-
-instance LowerBoundedField P.Float
-
-instance LowerBoundedField P.Double
-
-instance LowerBoundedField b => LowerBoundedField (a -> b) where
-  negInfinity _ = negInfinity
+-- | negative infinity
+negInfinity :: (Field a) => a
+negInfinity = negate infinity
 
 -- | Trigonometric Field
 class