+++ src/NumHask/Algebra/Metric.hs
@@ -14,6 +14,8 @@ module NumHask.Algebra.Metric
     polar,
     coord,
     Epsilon (..),
+    nearZero,
+    aboutEqual,
     (~=),
   )
 where
@@ -23,7 +25,7 @@ import Data.Int (Int16, Int32, Int64, Int8)
 import Data.Word (Word16, Word32, Word64, Word8)
 import GHC.Generics (Generic)
 import GHC.Natural (Natural (..))
-import NumHask.Algebra.Additive (Additive (zero), Subtractive (..))
+import NumHask.Algebra.Additive (Additive (zero), Subtractive (..), (-))
 import NumHask.Algebra.Lattice (MeetSemiLattice, meetLeq)
 import NumHask.Algebra.Module (MultiplicativeAction ((.*)))
 import NumHask.Algebra.Multiplicative (Multiplicative (one))
@@ -35,6 +37,11 @@ import Prelude hiding
     (-),
   )
 import qualified Prelude as P
+
+-- $setup
+--
+-- >>> :set -XRebindableSyntax
+-- >>> import NumHask.Prelude
 
 -- | 'signum' from base is not an operator name in numhask and is replaced by 'sign'.  Compare with 'Norm' where there is a change in codomain.
 --
@@ -249,15 +256,23 @@ class
   epsilon :: a
   epsilon = zero
 
-  nearZero :: a -> Bool
-  nearZero a = epsilon `meetLeq` a && epsilon `meetLeq` negate a
+-- | are we near zero?
+--
+-- >>> nearZero (epsilon :: Double)
+-- True
+nearZero :: (Epsilon a) => a -> Bool
+nearZero a = epsilon `meetLeq` a && epsilon `meetLeq` negate a
 
-  aboutEqual :: a -> a -> Bool
-  aboutEqual a b = nearZero $ a - b
+-- | Approximate equality
+aboutEqual :: (Epsilon a) => a -> a -> Bool
+aboutEqual a b = nearZero $ a - b
 
 infixl 4 ~=
 
--- | About equal.
+-- | About equal operator.
+--
+-- >>> (1.0 + epsilon) ~= (1.0 :: Double)
+-- True
 (~=) :: (Epsilon a) => a -> a -> Bool
 (~=) = aboutEqual