+++ src/NumHask/Algebra/Module.hs
@@ -5,26 +5,21 @@
 -- | Algebra for Modules
 module NumHask.Algebra.Module
   ( AdditiveAction (..),
+    (+.),
     SubtractiveAction (..),
+    (-.),
     MultiplicativeAction (..),
+    (*.),
     DivisiveAction (..),
+    (/.),
     Module,
   )
 where
 
-import NumHask.Algebra.Additive (Additive, Subtractive)
-import NumHask.Algebra.Multiplicative (Divisive, Multiplicative)
+import NumHask.Algebra.Additive (Additive, Subtractive, negate)
+import NumHask.Algebra.Multiplicative (Divisive, Multiplicative, recip)
 import NumHask.Algebra.Ring (Distributive)
-
--- $setup
---
--- >>> :set -XRebindableSyntax
--- >>> :set -XFlexibleContexts
--- >>> :set -XFlexibleInstances
--- >>> :set -XScopedTypeVariables
--- >>> :set -XMultiParamTypeClasses
--- >>> import NumHask.Prelude
--- >>> import Prelude (Int, fmap)
+import Prelude (flip)
 
 -- | Additive Action
 class
@@ -35,8 +30,13 @@ class
   infixl 6 .+
   (.+) :: a -> m -> m
 
-  infixl 6 +.
-  (+.) :: m -> a -> m
+infixl 6 +.
+
+-- | flipped additive action
+--
+-- > (+.) == flip (.+)
+(+.) :: (AdditiveAction m a) => m -> a -> m
+(+.) = flip (.+)
 
 -- | Subtractive Action
 class
@@ -47,8 +47,13 @@ class
   infixl 6 .-
   (.-) :: a -> m -> m
 
-  infixl 6 -.
-  (-.) :: m -> a -> m
+infixl 6 -.
+
+-- | right scalar subtraction
+--
+-- > (-.) == (+.) . negate
+(-.) :: (AdditiveAction m a, Subtractive a) => m -> a -> m
+a -. b = a +. negate b
 
 -- | Multiplicative Action
 class
@@ -58,8 +63,14 @@ class
   where
   infixl 7 .*
   (.*) :: a -> m -> m
-  infixl 7 *.
-  (*.) :: m -> a -> m
+
+infixl 7 *.
+
+-- | flipped multiplicative action
+--
+-- > (*.) == flip (.*)
+(*.) :: (MultiplicativeAction m a) => m -> a -> m
+(*.) = flip (.*)
 
 -- | Divisive Action
 class
@@ -69,8 +80,12 @@ class
   where
   infixl 7 ./
   (./) :: a -> m -> m
-  infixl 7 /.
-  (/.) :: m -> a -> m
+
+-- | right scalar division
+--
+-- > (/.) == (*.) . recip
+(/.) :: (MultiplicativeAction m a, Divisive a) => m -> a -> m
+a /. b = a *. recip b
 
 -- | A <https://en.wikipedia.org/wiki/Module_(mathematics) Module>
 --